Created extra source files for json.
