+++ engine/CMakeLists.txt
@@ -69,6 +69,8 @@ add_library(${PROJECT_NAME} STATIC
         src/core/ecs/system/script_ec_system.c
         src/core/ecs/system/sprite_rendering_ec_system.c
         src/core/json/json_file_loader.c
+        src/core/json/json_file_loader_scene.c
+        src/core/json/json_helper.c
         src/core/math/curve_float_manager.c
         )
